requirements.txt
amqp==2.6.1
asgiref==3.3.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
billiard==3.6.3.0
celery==4.4.7
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
constantly==15.1.0
cryptography==3.3.1
Django==3.1.4
django-celery-beat==2.1.0
django-filter==2.4.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
humanize==3.2.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
iniconfig==1.1.1
kombu==4.6.11
loguru==0.5.3
Markdown==3.3.3
mock==4.0.3
packaging==20.8
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
py==1.10.0
pycparser==2.20
PyHamcrest==2.0.2
pyparsing==2.4.7
pytest==6.2.1
pytest-mock==3.5.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
Twisted==20.3.0
txaio==20.12.1
urllib3==1.26.2
vine==1.3.0
wcwidth==0.2.5
websockets==8.1
zope.interface==5.2.0
